refactor: simplify bit reader and writer

+++ src/lib/Flood/IBitReader.cs
@@ -7,6 +7,6 @@ namespace Piot.Flood
 {
     public interface IBitReader
     {
-        public ulong ReadBits(int bitCount);
+        public uint ReadBits(int bitCount);
     }
 }
